trying to get this past compilation on macOS
