+++ Tests/VoxelsTests/PerformanceTests/SurfaceNetPerformanceTests.swift
@@ -3,7 +3,7 @@ import XCTest
 
 final class SurfaceNetPerformanceTests: XCTestCase {
     func testPerformanceInitialSurfaceNetPort() throws {
-        let samples = try VoxelTestHelpers.voxelArrayFromSphere()
+        let samples = try SampleMeshData.voxelArrayFromSphere()
 
         // 0.189 sec - original // initial 0.220
         // changing corners: 0.183, 0.184 // initial 0.216
@@ -12,6 +12,16 @@ final class SurfaceNetPerformanceTests: XCTestCase {
                 sdf: samples,
                 within: VoxelBounds(min: VoxelIndex(0, 0, 0), max: VoxelIndex(32, 32, 32))
             )
+        }
+    }
+
+    func testPerformanceRemadeSurfaceNetPort() throws {
+        let samples = try SampleMeshData.voxelArrayFromSphere()
+
+        let thing = SurfaceNetRenderer()
+        // 0.163, 0.165 sec - so very much on par with the original
+        measure {
+            _ = try! thing.render(voxelData: samples, scale: .init())
         }
     }
