+++ Tests/VoxelsTests/RendererTests/OriginalSurfaceNetTests.swift
@@ -3,7 +3,7 @@ import XCTest
 
 final class OriginalSurfaceNetTests: XCTestCase {
     func testVoxelArraySampling() throws {
-        let samples = try VoxelTestHelpers.voxelArrayFromSphere()
+        let samples = try SampleMeshData.voxelArrayFromSphere()
         let insides = samples.filter { val in
             val < 0
         }
@@ -11,7 +11,7 @@ final class OriginalSurfaceNetTests: XCTestCase {
     }
 
     func testExample() throws {
-        let samples = try VoxelTestHelpers.voxelArrayFromSphere()
+        let samples = try SampleMeshData.voxelArrayFromSphere()
         let buffer = try VoxelMeshRenderer.surfaceNetMesh(
             sdf: samples,
             within: VoxelBounds(min: VoxelIndex(0, 0, 0), max: VoxelIndex(32, 32, 32))
